extract validation logic from Developer API into pkg/validation
Extracted v0/devServer.Validate into pkg/validationfile.Validate, and
reused that logic in both devServer and CLI validate command. There are
several consequences:

- new API method validationfile/RequestFromFile. Exposes v0.ValidateRequest
- new API method validationfile/Validate. Exposes v0.ValidateRequest and *v0.ValidateResponse
- former v0/rewriteGraphError is used by the validation logic to construct the DeveloperError,
which is useful for error reporting in the CLI. It was tricky to move around because it is used by
various components, like DevContext and devServer. As a consequence I exported it,
which I wasn't too fond of, as it seemed rather implementation specific
- One important change is that former v0/rewriteGraphError no longer invokes rewriteACLError (method
used to wrap errors with gRPC error codes). The assumption I made (I couldn't validate) was that it's not
needed because v0.aclServer is already doing that, and presumably the error would have been already
 wrapped when it is returned to the devServer (e.g. after a Check command).
- DevContext is used by both CLI and devServer to prepare a Dispatcher that uses the
schema and relationships provided by the user in the YAML file. Duplicating all that logic
in order to decouple CLI from DevContext seemed counter-productive, and DevContext is
already exposed as internal API anyway. However, DevContext.dispose became exported in order
to encapsulate the lifecycle of the dispatcher and in memory database. That seemed like a reasonable
tradeoff.
- I duplicated maxDepth across validationfile and devserver. I had explored also moving EditCheck
to validationfile package, but I don't have a use-case right now so I left it there.
